// src/app/models/orderDetails.ts
export class OrderDetails {
  cartTotal: number = 0;
  firstName: string = '';
  lastName: string = '';
  email: string = '';
  addressLine1: string = '';
  addressLine2: string = '';
  city: string = '';
  state: string = '';
  zipCode: string = '';
  paymentAddress: string = '';

  constructor() {}
}